feat: refactor CSV download functionality into a separate component
